Update and rename YYYY-MM-DD-title.md to 2021-08-27-title.md
First, I want to tell you why I decided to start training on GitHub. 

You can talk for a long time about the age of digital technology and the huge influence of programming on modern human life. It is all understandable. It makes no sense to tell about the demand for specialists in the field of information technology. And if we add free work schedule, absence of attachment to the place of work and high wages to the demand, then this specialty is a dream!)

That's the dream I want to start with))